fixed fakemodule under python3
